Extract feature specific code from main feature router file (content.js) into feature file (preserve-text-messages.js)
Extract feature specific code from main feature router file (content.js) into feature file (preserve-text-messages.js) — https://github.com/sindresorhus/refined-twitter/pull/83#issuecomment-392126117
